# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.Game2048SaveRequest:
    properties:
      score:
        minimum: 0
        type: integer
    required:
    - score
    type: object
...
  controllers.updateRate:
    properties:
      from_currency:
        type: string
      rate:
        type: number
      to_currency:
        type: string
    type: object
  controllers.updateRateResponse:
    properties:
      date:
        type: string
      from_currency:
        type: string
      id:
        type: integer
      rate:
        type: number
      to_currency:
        type: string
    type: object
...
info:
  contact: {}
  description: Go-Web综合性应用API接口文档
  title: Go-Web项目 API
  version: 0.0.3
paths:
  /ad:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 示例：{\"author_url\":\"https://...\"}
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取作者博客广告
      tags:
      - Me
...
  /exchangeRates:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: 示例：{\"username\":\"alice\"}
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取当前汇率数据
      tags:
      - Me
  /exchangeRates/{id}:
    delete:
      consumes:
      - application/json
      description: 根据ID删除指定的汇率记录
      parameters:
      - description: 汇率记录ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功删除
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 无效的ID
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 汇率记录不存在
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 删除汇率记录
      tags:
      - ExchangeRate
    put:
      consumes:
      - application/json
      description: 根据ID更新指定的汇率记录信息
      parameters:
      - description: 汇率记录ID
        in: path
        name: id
        required: true
        type: integer
      - description: 更新汇率请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/controllers.updateRate'
      produces:
      - application/json
      responses:
        "200":
          description: 成功更新
          schema:
            $ref: '#/definitions/controllers.updateRateResponse'
        "400":
          description: 无效的请求参数
          schema:
            additionalProperties: true
            type: object
        "401":
          description: 未授权
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器内部错误
          schema:
            additionalProperties: true
            type: object
      security:
      - ApiKeyAuth: []
      summary: 更新汇率记录
      tags:
      - ExchangeRate
...
  /game/2048/save:
    post:
      consumes:
      - application/json
      description: 保存用户的2048游戏分数到数据库
      parameters:
      - description: 游戏分数
        in: body
        name: score
        required: true
        schema:
          $ref: '#/definitions/controllers.Game2048SaveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: 保存2048游戏分数
      tags:
      - Game
...
  /me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: 示例：{\"username\":\"alice\"}
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 获取当前用户名
      tags:
      - Me
...
  /rmb-top10:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              additionalProperties: true
              type: object
            type: array
      security:
      - Bearer: []
      summary: 读取当前人民币对Top10地区的汇率快照（来自 Redis）
      tags:
      - Exchange
  /rmb-top10/refresh:
    post:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - Bearer: []
      summary: 手动刷新人民币 Top10 汇率（写入 Redis，TTL=24h）
      tags:
      - Exchange
...
swagger: "2.0"